Added FPS counter (thanks nova) and optimized how Notes are actually added and rendered. So optimized
Before with notes, I was like "Well yah if they aren't being rendered (invisible) and not updating well then it's fine just to precache all the notes in a Node2D" and that made the FPS go to 150.

This new change with an RTX 3060 runs at 4400 fps with no notes on screen, and 2200fps with a basic chart loaded.